Define gates for authorization, protect actions for Attendee and Event

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\AttendeeResource;
 use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -16,7 +17,7 @@ class AttendeeController extends Controller
 
     public function __construct()
     {
-        $this->middleware('auth:sanctum')->except('index', 'show');
+        $this->middleware('auth:sanctum')->except('index', 'show', 'update');
     }
 
     /**
@@ -60,8 +61,12 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        if (Gate::denies('delete-attendee', [$event, $attendee])) {
+            abort(403, 'This action is not authorized');
+        }
+
         $attendee->delete();
 
         return response(status: 204);
